+++ action_server/src/sema4ai/action_server/_cli_helpers.py
@@ -29,6 +29,59 @@ def add_verbose_args(parser, defaults):
     )
 
 
+def add_login_args(parser):
+    parser.add_argument(
+        "-a",
+        "--access-credentials",
+        type=str,
+        help="Control Room access credentials",
+        required=False,
+    )
+    parser.add_argument(
+        "--hostname", type=str, help="Hostname for the Control Room URL", required=False
+    )
+
+
+def add_organization_args(parser):
+    parser.add_argument(
+        "-o",
+        "--organization-id",
+        type=str,
+        help="Control Room organization ID",
+        required=True,
+    )
+
+
+def add_push_package_args(parser):
+    parser.add_argument(
+        "-p",
+        "--package-path",
+        type=str,
+        help="Path to the uploaded package",
+        required=True,
+    )
+
+
+def add_package_args(parser):
+    parser.add_argument(
+        "-i",
+        "--package-id",
+        type=str,
+        help="Control Room action package ID",
+        required=True,
+    )
+
+
+def add_publish_args(parser):
+    parser.add_argument(
+        "-c",
+        "--change-log",
+        type=str,
+        help="Changelog string for published package",
+        required=True,
+    )
+
+
 def add_json_output_args(parser):
     parser.add_argument(
         "--json",